refactor: address review comments - Configure ESLint to ignore underscore-prefixed vars (varsIgnorePattern: '^_') - Remove all inline `eslint-disable` comments from the codebase

+++ backend/src/films/entities/schedule.entity.ts
@@ -18,7 +18,7 @@ export class Schedule {
   @Column({ type: 'int' })
   seats: number;
 
-  @Column({ type: 'int' })
+  @Column({ type: 'float' })
   price: number;
 
   @Column({ type: 'simple-array', default: '' })
